LPD-69571 Delete obsolete poshi tests
We are deleting these tests because:

1. There's nothing like a RLT infrastructure. Obviously we support RTL in
   CSS files, but it's mainly a build time thing plus small pieces of code
   that select the correct file to use.
2. The tests are very simple and don't test RTL at all.
3. The tests are quite unclear and it's not easy to tell why they are doing
   what they are doing.
4. We asked Platform Experience if they wanted to own any of these tests
   and they said more or less the same.

We may create more tests to check RTL in the future, but they will be
written in Playwright and do something meaningful. Meanwhile, we will
delete this two to avoid noise in the daily checks.

Note that even one of the two tests was quarantined long ago and was kept
like that for several months.