LPD-61657 | LPD-61658 Implement a ModelResourcePermissionConfigurator to check permission when an user triggers the ObjectField download url.
Extract a method to handle when files are exclusive from ObjectDefinition side. It has to ensure the download permission is always verified, if not, don't grant it. Use _log.error() instead of throwing an exception because it would affect the request behavior(guest users would not go to 404 page if exception is thrown).